Set up a Jenkins pipeline to test PRs and deploy in production
A stub Jenkinsfile, which defines a pipeline to test PRs and deploy
software in production, has been committed.

The pipeline has been tested on our internal Jenkins, and is based
on a few naming conventions for epfl-scitas/spack and
epfl-scitas/spack-packagelist (for each PR there must be 2 branches
with the same name in both repositories).

The stub 'echo' directives will be substituted later, as soon as agents
will be available.