Added scripts to test a PR (#41)
Spack is checked out in a temporary folder, where we try to build the list of software that is not yet installed in production. The entire configuration is copied into the temporary Spack instance, with the exception of config.yaml (to avoid installing directly into the production folders). Compilers are taken directly from the produciton environment.