Remove CLASSPATH from the list of variables that ends in Intel's module
CLASSPATH apparently causes each JVM instance to open all the files
that are listed in the variable. This triggers a yet to be identified
bug (probably related to GPFS) that blocks the node in a sort
of invalid state and prevents the backup of the filesystem where
the process keeping the files open started.

See HPC-1410 fo an example of this bug.