Separate deployment in production from PRs in Jenkinsfile
The Jenkinsfile has been separated into two parts: one deals with the
deployment of software in production, the other with testing new
software in PRs.

The two new Jenkinsfile are kept in separate folders, along with the
shell scripts needed to run them. All references to paien have been
modified to refer to humagne.